+++ core/src/main/java/org/calyxos/seedvault/core/backends/saf/SafProperties.kt
@@ -7,8 +7,11 @@ package org.calyxos.seedvault.core.backends.saf
 
 import android.content.Context
 import android.net.Uri
+import android.provider.DocumentsContract
 import android.provider.DocumentsContract.Root.COLUMN_ROOT_ID
+import android.util.Log
 import androidx.annotation.WorkerThread
+import androidx.core.database.getStringOrNull
 import androidx.documentfile.provider.DocumentFile
 import org.calyxos.seedvault.core.backends.BackendProperties
 
@@ -37,6 +40,37 @@ public data class SafProperties(
      */
     @WorkerThread
     override fun isUnavailableUsb(context: Context): Boolean {
-        return isUsb && !getDocumentFile(context).isDirectory
+        if (!isUsb) return false
+        return if (rootId == null) { // fallback for when we didn't store rootId
+            // the document file is not a directory
+            !getDocumentFile(context).isDirectory
+        } else {
+            // retry root check due to SAF bug
+            for (i in 1..3) {
+                try {
+                    // if root isn't present, the usb storage is unavailable
+                    return !isRootIdPresent(context)
+                } catch (e: Exception) {
+                    Log.e("SafProperties", "Error getting root ($i): ", e)
+                    continue
+                }
+            }
+            return true // we had three exceptions, so is unavailable
+        }
+    }
+
+    private fun isRootIdPresent(context: Context): Boolean {
+        val rootUri = DocumentsContract.buildRootsUri(uri.authority)
+        val projection = arrayOf(COLUMN_ROOT_ID)
+        return context.contentResolver.query(
+            rootUri, projection, "$COLUMN_ROOT_ID = ?", arrayOf(rootId), null
+        )?.use { c ->
+            // the selection per root ID doesn't work, so we need to look at cursor rows
+            while (c.moveToNext()) {
+                val str = c.getStringOrNull(c.getColumnIndex(COLUMN_ROOT_ID))
+                if (str == rootId) return true
+            }
+            return false
+        } ?: throw NullPointerException()
     }
 }